requirements.txt
requests==2.32.0
azure-ai-inference==1.0.0b9
